# Import: Cleanup and optimize etcpak compression method 