--- a/modules/etcpak/register_types.cpp
+++ b/modules/etcpak/register_types.cpp
@@ -30,7 +30,7 @@
 
 #include "register_types.h"
 
-#include "image_etcpak.h"
+#include "image_compress_etcpak.h"
 
 void register_etcpak_types() {
 	Image::_image_compress_etc1_func = _compress_etc1;