--- a/modules/squish/register_types.cpp
+++ b/modules/squish/register_types.cpp
@@ -29,7 +29,8 @@
 /*************************************************************************/
 
 #include "register_types.h"
-#include "image_compress_squish.h"
+
+#include "image_decompress_squish.h"
 
 void register_squish_types() {
 	Image::_image_decompress_bc = image_decompress_squish;