Avoid unnecessary allocation of temporary buffers for each mip, and redundant
Image `create` where we can assign the pointer directly.
Also renames variable and reorders code for clarity.

Clarify that squish is now only used for decompression.

Documented which formats can be decompressed in Image.

Follow-up to #47370.

## Notes

### Lossy quality

etcpak doesn't provide any options for configuring the quality. Reading the [upstream README](https://github.com/wolfpld/etcpak), it says:

> etcpak is an extremely fast Ericsson Texture Compression and S3 Texture Compression (DXT1/DXT5) utility. Currently it's best suited for rapid assets preparation during development, when graphics quality is not a concern, but it's also used in production builds of applications used by millions of people.

i.e. it's not geared towards perfect quality, but then that's also aligned with our default lossy import with quality 0.7. That means that if some users eventually want higher quality, we might need to provide another way to encode their assets (possibly via a thirdparty plugin).

But it also states:

> [Why there's no image quality metrics? / Quality comparison.](https://i.imgur.com/FxlmUOF.png)

And the image comparison does look fairly good to my untrained eye.

So now that we import ETC/ETC2 and S3TC through an importer that has no configurable quality settings, maybe we should consider removing the `lossy_quality` option?

The PVRTC importer we use also doesn't handle `lossy_quality`, so only cvtt handles it for BPTC.

### Decompression

etcpak does provide support for decode ETC, ETC2, DXT1, and DXT5.

It's not part of the minimal `Process*` code we extracted from the library, but part of the `BlockData.cpp` which depends on etcpak's own Bitmap representation and everything else: task scheduler, modified libpng, etc.

https://github.com/wolfpld/etcpak/blob/1dc7b300453ae3d20cc43d906c89bcfcd8854cd3/BlockData.cpp#L694-L711

We could likely extract only the relevant `Decode*Part` methods and hook them with our own `Image` bitmap representation: https://github.com/wolfpld/etcpak/blob/1dc7b300453ae3d20cc43d906c89bcfcd8854cd3/BlockData.cpp#L723-L857

That would require either some copy-paste work which makes it harder to sync with upstream, or splitting the upstream `BlockData.cpp` into two files (one part with only the byte decoding, and another part with the higher level Bitmap stuff) and have them agree to merge it.

That would allow use to drop Squish (used to decode DXT) and add decoding support for ETC1/ETC2.