@akien-mga 2021-04-16
I went back to a safer approach closer to the original code (each mip gets written to a temporary buffer and `memcpy`'ed) as I wasn't sure how to get the full size needed for the new image to pre-allocate it.

Doing so I found that I was mixing metrics of the base and dest images which could have explained the crash I was getting with the old method. I might poke at it further but in the meantime this version should be somewhat safer. It imports all textures from the TPS demo fine.