@akien-mga 2021-04-16
> I went back to a safer approach closer to the original code (each mip gets written to a temporary buffer and `memcpy`'ed) as I wasn't sure how to get the full size needed for the new image to pre-allocate it.

In the end I managed to find a way to avoid the temporary buffers. As in the original code, we `create` the new `Image` beforehand, so the data buffer is pre-allocated and we just need to write to it directly.
@fire told me the initial implementation was done with intermediate arrays so that it could run multi-threaded and be joined in the end, but since we'll now have threaded import of multiple resources in parallel with #47343, this is likely not needed.