@akien-mga 2021-04-16
Tested before/after with an optimized build to ensure that this PR actually helps or doesn't reduce performance.

Importing all `.stex` for TPS demo with an optimize build (`production=yes target=release_debug`).

- Before: 0m49,701s 0m50,410s 0m50,549s
- After: 0m58,678s 0m56,395s 0m57,866s

So it seems I'm actually losing some performance here, surprisingly. I'll have to look into it further.