@akien-mga 2021-04-16
Cleaned up further with advice from @reduz so we no longer need to create an intermediate `Image` to pre-allocate the right data. `Image` has all the relevant static methods to get the image size and mipmap offset and dimensions for a given width/height/format.

With these changes, I get the same performance before this PR and after on my hardware (~49-50 s). Hard to say if there's an actual speed gain as numbers fluctuate significantly depending on the start CPU load and heat on my laptop.

For the reference, to test I do:
1. Import TPS demo fully one time, exit
2. `rm .godot/imported/*.stex -f`
3. `time godot -e` and I manually abort (Ctrl+C) once I reach 100%
4. Repeat 2. and 3. with the other version being tested